+++ src/main/java/frc/robot/commands/driving/DirectDriveCommand.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.commands.driving;
 
 import java.util.function.DoubleSupplier;
