+++ src/main/java/frc/robot/commands/intake/IntakeCommand.java
@@ -5,28 +5,31 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands;
+package frc.robot.commands.intake;
+
+import java.util.function.DoubleSupplier;
+
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
-/*----------------------------------------------------------------------------*/
-/* Copyright (c) 2019 FIRST. All Rights Reserved.                             */
-/* Open Source Software - may be modified and shared by FRC teams. The code   */
-/* must be accompanied by the FIRST BSD license file in the root directory of */
-/* the project.                                                               */
-/*----------------------------------------------------------------------------*/
-import frc.robot.subsystems.ClimberSubsystem;
+import frc.robot.subsystems.IntakeSubsystem;
 
+public class IntakeCommand extends CommandBase {
 
-public class TurnOffClimberCommand extends CommandBase {
+private final DoubleSupplier spin; 
   /**
-   * Creates a new ClimberCommand.
+   * Creates a new IntakeCommand.
    */
-  ClimberSubsystem climberSubsystem;
-  public TurnOffClimberCommand(ClimberSubsystem subsystem) {
-    // Use addRequirements() here to declare subsystem dependencies.
-    climberSubsystem = subsystem;
-    addRequirements(climberSubsystem);
-  }
+  IntakeSubsystem intakeSubsystem;
+  public IntakeCommand(IntakeSubsystem subsystem, DoubleSupplier spin) {
+    intakeSubsystem = subsystem;
+    this.spin = spin;
+    addRequirements(intakeSubsystem);
+
+
+
+
+}
+
 
   // Called when the command is initially scheduled.
   @Override
@@ -36,14 +39,13 @@ public class TurnOffClimberCommand extends CommandBase {
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    climberSubsystem.disable();
-    
+    intakeSubsystem.enable(spin.getAsDouble()); 
+
   }
 
   // Called once the command ends or is interrupted.
   @Override
   public void end(boolean interrupted) {
-
   }
 
   // Returns true when the command should end.
@@ -52,4 +54,3 @@ public class TurnOffClimberCommand extends CommandBase {
     return false;
   }
 }
-
